* reasonably large performance boost for ai, no longer searching lists to find threat for spell, its stored in spell struct * crash fixes
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@3336 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/WorldSession.cpp
@@ -220,7 +220,7 @@ int WorldSession::Update(uint32 InstanceID)
 			_socket = NULL;
 		}
 
-		m_lastPing = UNIXTIME;		// Prevent calling this code over and over.
+		m_lastPing = (uint32)UNIXTIME;		// Prevent calling this code over and over.
 		if(!_logoutTime)
 			_logoutTime = m_currMsTime + PLAYER_LOGOUT_DELAY;
 	}
